requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docker==7.0.0
exceptiongroup==1.2.2
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hubspot-api-client==11.1.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.20.1
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
